In SharedData, save Pair<Class<?>,?> where ? is the object, instead of JSON serialization.
Now we are using a static map to save data, and data can be objects. In previous commits, SharedPreferences had to be string, that is why data were JSON serialized before saving: it required a computational effort (JSON serialization for saving and JSON deserialization for retrieval) which is not needed anymore.